chores(testing): cleanup unit tests
