--- src/entities/json/json.c
@@ -1,5 +1,0 @@
-//
-// Created by Paul Gaudeaux on 25/09/2023.
-//
-
-#include "json.h"
