--- src/entities/json/json_test.c
@@ -1,38 +1,0 @@
-#include <stdio.h>
-#include "../../minunit/minunit.h"
-
-int tests_run = 0;
-
-int foo = 7;
-int bar = 4;
-
-static char * test_foo() {
-    mu_assert("error, foo != 7", foo == 7);
-    return 0;
-}
-
-static char * test_bar() {
-    mu_assert("error, bar != 5", bar == 4);
-    return 0;
-}
-
-static char * all_tests() {
-    mu_run_test(test_foo);
-    mu_run_test(test_bar);
-    return 0;
-}
-
-int json_test() {
-    char *result = all_tests();
-    if (result != 0) {
-        printf("%s\n", result);
-    }
-    else {
-        printf("ALL TESTS PASSED\n");
-    }
-    printf("Tests run: %d\n", tests_run);
-
-    return result != 0;
-}
-
-
