+++ src/storage/json/json.h
@@ -1,7 +1,3 @@
-//
-// Created by Paul Gaudeaux on 25/09/2023.
-//
-
 #ifndef DOOM_DEPTH_C_JSON_H
 #define DOOM_DEPTH_C_JSON_H
