+++ test.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include "src/minunit/minunit.h"
-#include "src/entities/json/json.h"
+#include "src/storage/json/json.h"
 
 int main(int argc, char **argv) {
     return json_test();
